xtensa/esp32: Fix code nxstyle issue
