+++ arch/xtensa/src/esp32/Make.defs
@@ -78,10 +78,10 @@ CHIP_CSRCS += esp32_gpio.c esp32_intdecode.c esp32_irq.c esp32_region.c
 CHIP_CSRCS += esp32_user.c
 CHIP_CSRCS += esp32_dma.c
 
-ifneq ($(CONFIG_SCHED_TICKLESS),y)
-CHIP_CSRCS += esp32_timerisr.c
-else
+ifeq ($(CONFIG_SCHED_TICKLESS),y)
 CHIP_CSRCS += esp32_tickless.c
+else
+CHIP_CSRCS += esp32_timerisr.c
 endif
 
 ifeq ($(CONFIG_PM),y)
