+++ boards/xtensa/esp32/esp32-devkitc/configs/autopm/defconfig
@@ -63,7 +63,6 @@ CONFIG_SCHED_LPWORK=y
 CONFIG_SCHED_WAITPID=y
 CONFIG_SDCLONE_DISABLE=y
 CONFIG_SIG_DEFAULT=y
-CONFIG_SPI=y
 CONFIG_SPIFFS_NAME_MAX=48
 CONFIG_START_DAY=6
 CONFIG_START_MONTH=12
