+++ boards/xtensa/esp32/esp32-ethernet-kit/configs/autopm/defconfig
@@ -62,7 +62,6 @@ CONFIG_SCHED_LPWORK=y
 CONFIG_SCHED_WAITPID=y
 CONFIG_SDCLONE_DISABLE=y
 CONFIG_SIG_DEFAULT=y
-CONFIG_SPI=y
 CONFIG_SPIFFS_NAME_MAX=48
 CONFIG_START_DAY=6
 CONFIG_START_MONTH=12
